pyproject.toml
[tool.poetry]
name = "soil-api"
version = "0.1.1"
description = "API to manage data for SOIL lab"
authors = ["Evan Thomas <evan.thomas@epfl.ch>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
geoalchemy2 = "^0.14.2"
httpx = "^0.25.1"
asyncpg = "^0.29.0"
shapely = "^2.0.2"
sqlalchemy-utils = "^0.41.1"
alembic = "^1.12.1"
pydantic-geojson = "^0.1.1"
numpy = "^1.26.1"
geos = "^0.2.3"
gpxpy = "^1.6.1"
pydantic-settings = "^2.2.1"
pyproj = "^3.6.1"
secretcolors = "^1.2.6"
alembic-postgresql-enum = "^1.2.0"
tenacity = "^8.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-watcher = "^0.3.4"
pytest-asyncio = "^0.21.1"
aiosqlite = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
